Don't override ternProjectDir if already set in vimrc.
The function term#Enable overrides the value of ternProjectDir. In some
cases we would need to allow to customize this value, for instance
working with multiple projects that have similar conditions where we
don't want to create a new .term-project file for every project.

With this change, we can define a custom .term-project file in .vimrc that
will be applied to every project.