Add option for showing loclist after :TernRefs
Only open the location list if g:tern_show_loc_after_refs is set to 1.
The location list is always populated and if g:tern_show_loc_after_refs
is not 1 the location list entry under the cursor will be selected for
easy navigation with :lprev and :lnext.

Note that the behaviour of :TernRename is changed to be consistent with
the new functionality of :TernRefs, which means that the location list
will be populated even if g:tern_show_loc_after_rename is not 1.